requirements.txt
pandas
numpy
matplotlib
scipy
pingouin
scikit-posthocs
seaborn
openpyxl
pyyaml